+++ ocelot/util/interface/KernelExtractor.h
@@ -111,7 +111,7 @@ namespace util {
 		CUresult cuDeviceGetName(char *name, int len, CUdevice dev);
 		CUresult cuDeviceComputeCapability(int *major, int *minor, 
 			CUdevice dev);
-		CUresult cuDeviceTotalMem(unsigned int *bytes, CUdevice dev);
+		CUresult cuDeviceTotalMem(size_t *bytes, CUdevice dev);
 		CUresult cuDeviceGetProperties(CUdevprop *prop, 
 			CUdevice dev);
 		CUresult cuDeviceGetAttribute(int *pi, 
@@ -151,7 +151,7 @@ namespace util {
 		CUresult cuModuleGetFunction(CUfunction *hfunc, 
 			CUmodule hmod, const char *name);
 		CUresult cuModuleGetGlobal(CUdeviceptr *dptr, 
-			unsigned int *bytes, CUmodule hmod, const char *name);
+			size_t *bytes, CUmodule hmod, const char *name);
 		CUresult cuModuleGetTexRef(CUtexref *pTexRef, CUmodule hmod, 
 			const char *name);
 
@@ -161,20 +161,19 @@ namespace util {
 		**
 		***********************************/
 
-		CUresult cuMemGetInfo(unsigned int *free, 
-			unsigned int *total);
+		CUresult cuMemGetInfo(size_t *free, size_t *total);
 
 		CUresult cuMemAlloc( CUdeviceptr *dptr, 
 			unsigned int bytesize);
 		CUresult cuMemAllocPitch( CUdeviceptr *dptr, 
-					          unsigned int *pPitch,
+					          size_t *pPitch,
 					          unsigned int WidthInBytes, 
 					          unsigned int Height, 
 					          unsigned int ElementSizeBytes
 					         );
 		CUresult cuMemFree(CUdeviceptr dptr);
 		CUresult cuMemGetAddressRange( CUdeviceptr *pbase, 
-			unsigned int *psize, CUdeviceptr dptr );
+			size_t *psize, CUdeviceptr dptr );
 
 		CUresult cuMemAllocHost(void **pp, unsigned int bytesize);
 		CUresult cuMemFreeHost(void *p);
@@ -337,8 +336,8 @@ namespace util {
 
 		CUresult cuTexRefSetArray( CUtexref hTexRef, CUarray hArray, 
 			unsigned int Flags );
-		CUresult cuTexRefSetAddress( unsigned int *ByteOffset, 
-			CUtexref hTexRef, CUdeviceptr dptr, unsigned int bytes );
+		CUresult cuTexRefSetAddress( size_t *ByteOffset, 
+			CUtexref hTexRef, CUdeviceptr dptr, size_t bytes );
 		CUresult cuTexRefSetAddress2D( CUtexref hTexRef, 
 			const CUDA_ARRAY_DESCRIPTOR *desc, CUdeviceptr dptr, 
 			unsigned int Pitch);
@@ -429,7 +428,7 @@ namespace util {
 			CUarray *pArray, CUgraphicsResource resource, 
 			unsigned int arrayIndex, unsigned int mipLevel );
 		CUresult cuGraphicsResourceGetMappedPointer(
-			CUdeviceptr *pDevPtr, unsigned int *pSize, 
+			CUdeviceptr *pDevPtr, size_t *pSize, 
 			CUgraphicsResource resource );
 		CUresult cuGraphicsResourceSetMapFlags(
 			CUgraphicsResource resource, unsigned int flags );
